Allow setting pack path of managed plugins
The pack directory used for installing plugins is hard coded and cannot
be changed. This is an issue in some situations. For example, if a new
configuration is being created or tested, plugins already installed to
the default packpath locations may be clobbered; or if a different
non-standard location is already configured by the user before using
paq, this will surprise and/or confuse the user or even lead to the
plugins managed by paq to not be loaded by nvim.

This change exports a new function from the `paq-nvim` module called
`setup_path`. This new function simply allows setting the internal path
used to install plugins to a different location.